requirements.txt
pyserial==3.5
redis==5.0.8
python-dotenv==1.0.1
